requirements.txt
attrs==18.2.0
Click==7.0
docutils==0.14
lockfile==0.12.2
luigi==2.8.3
oradump==1.0
psycopg2==2.7.7
python-daemon==2.1.2
python-dateutil==2.7.5
python-dotenv==0.10.0
six==1.12.0
tornado==4.5.3
